Extract Sidebar components
